Add pagination for products with combination without quantities
